feat: create upper level EthChain struct
